Applied low coupling - high cohesion by dividing project into several solutions
