+++ Backend.Core/GlobalUsings.cs
@@ -1,0 +1,2 @@
+﻿global using Backend.Core.Interfaces;
+global using Backend.Core.Models;
