+++ Backend.Core/Interfaces/IBudget.cs
@@ -1,0 +1,8 @@
+﻿namespace Backend.Core.Interfaces;
+
+public interface IBudget
+{
+    public Category Income { get; set; }
+    public string? Title { get; set; }
+    public List<Category> Expenses { get; set; }
+}
