+++ Backend.Core/Interfaces/IBudgetFiller.cs
@@ -1,0 +1,9 @@
+﻿namespace Backend.Core.Interfaces;
+
+public interface IBudgetFiller
+{
+    public Budget? FetchLargeBudget();
+    public Budget? FetchMediumBudget();
+    public Budget? FetchSmallBudget();
+}
+
