+++ Backend.Core/Interfaces/IBudgetManager.cs
@@ -1,0 +1,13 @@
+﻿namespace Backend.Core.Interfaces;
+
+public interface IBudgetManager
+{
+    public Budget? SmallBudget { get; set; }
+    public Budget? MediumBudget { get; set; }
+    public Budget? LargeBudget { get; set; }
+
+    public Budget BudgetChecker(Budget budget);
+
+    public bool BudgetIsValid(Budget budget);        
+
+}
