+++ Backend.Core/Interfaces/ICategory.cs
@@ -1,0 +1,9 @@
+﻿namespace Backend.Core.Interfaces;
+
+public interface ICategory
+{
+    public int Id { get; set; }
+    public string? Name { get; set; }
+    public float TotalAmount { get; set; }
+    public List<Item> Items { get; set; }
+}
