+++ Backend.Core/Interfaces/IItem.cs
@@ -1,0 +1,8 @@
+﻿namespace Backend.Core.Interfaces;
+
+public interface IItem
+{
+    public int Id { get; set; }
+    public string? Name { get; set; }
+    public float Amount { get; set; }
+}
