+++ Backend.Core/Interfaces/IItemManager.cs
@@ -1,0 +1,8 @@
+﻿namespace Backend.Core.Interfaces;
+
+public interface IItemManager
+{
+    public bool CheckIfItemsAreValidInBudget(Budget budget);
+    public bool CheckValidItem(Item item);
+
+}
