+++ Backend.Core/Models/Budget.cs
@@ -1,0 +1,9 @@
+﻿namespace Backend.Core.Models;
+
+public class Budget : IBudget
+{
+    public Category Income { get; set; } = new Category();
+    public string? Title { get; set; }
+    public List<Category> Expenses { get; set; } = new List<Category>();
+    
+}
