+++ Backend.Test/BudgetTests.cs
@@ -1,73 +1,69 @@
-using Backend.DAL;
-using Backend.Models;
+namespace Backend.UnitTest;
 
-namespace Backend.Test
+public class BudgetTests : IClassFixture<BudgetFixture>
 {
-    public class BudgetTests : IClassFixture<BudgetFixture>
+    private BudgetManager _sut;
+    private Budget? _budget;
+    public BudgetTests(BudgetFixture budgetFixture)
     {
-        private BudgetManager _sut;
-        private Budget? _budget;
-        public BudgetTests(BudgetFixture budgetFixture)
-        {
-            _sut = budgetFixture.BudgetManager; 
-            _budget = budgetFixture.Budget;
-        }
+        _sut = budgetFixture.BudgetManager; 
+        _budget = budgetFixture.Budget;
+    }
 
 
-        [Fact]
-        public void BudgetReturnsValid()
-        {
-            // Arrange
+    [Fact]
+    public void BudgetReturnsValid()
+    {
+        // Arrange
 
-            // Act
-            var actual = _sut.BudgetIsValid(_budget);
+        // Act
+        var actual = _sut.BudgetIsValid(_budget);
 
-            // Assert
-            Assert.True(actual);
-        }
+        // Assert
+        Assert.True(actual);
+    }
 
-        [Fact]
-        public void BudgetsGetInitializedWhenCalledOnBudgetManager()
-        {
-            // Arrange
-            // Act
-            // Assert
-            Assert.NotNull(_sut.SmallBudget);
-            Assert.NotNull(_sut.MediumBudget);
-            Assert.NotNull(_sut.LargeBudget);
-        }
+    [Fact]
+    public void BudgetsGetInitializedWhenCalledOnBudgetManager()
+    {
+        // Arrange
+        // Act
+        // Assert
+        Assert.NotNull(_sut.SmallBudget);
+        Assert.NotNull(_sut.MediumBudget);
+        Assert.NotNull(_sut.LargeBudget);
+    }
 
-        [Fact]
-        public void LargeBudgetGets8CategoriesAsDefault()
-        {
-            // Arrange
-            var expected = 8;
+    [Fact]
+    public void LargeBudgetGets8CategoriesAsDefault()
+    {
+        // Arrange
+        var expected = 8;
 
-            //Act
+        //Act
 
-            // Assert
-            Assert.Equal(expected, _sut.LargeBudget?.Expenses.Count);
-        }
+        // Assert
+        Assert.Equal(expected, _sut.LargeBudget?.Expenses.Count);
+    }
 
-        [Fact]
-        public void MediumBudgetGets7CategoriesAsDefault()
-        {
-            // Arrange
-            var expected = 7;
+    [Fact]
+    public void MediumBudgetGets7CategoriesAsDefault()
+    {
+        // Arrange
+        var expected = 7;
 
-            // Act
-            // Assert
-            Assert.Equal(expected, _sut.MediumBudget?.Expenses.Count);
-        }
+        // Act
+        // Assert
+        Assert.Equal(expected, _sut.MediumBudget?.Expenses.Count);
+    }
 
-        [Fact]
-        public void SmallBudgetGets6CategoriesAsDefault()
-        {
-            // Arrange
-            var expected = 6;
+    [Fact]
+    public void SmallBudgetGets6CategoriesAsDefault()
+    {
+        // Arrange
+        var expected = 6;
 
-            // Act
-            Assert.Equal(expected, _sut.SmallBudget?.Expenses.Count);
-        }
+        // Act
+        Assert.Equal(expected, _sut.SmallBudget?.Expenses.Count);
     }
 }
