--- Backend.Test/Data/TheroyData.cs
@@ -1,29 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Backend.Test.Data
-{
-    public abstract class TheoryData : IEnumerable<object[]>
-    {
-        readonly List<object[]> data = new List<object[]>();
-
-        protected void AddRow(params object[] values)
-        {
-            data.Add(values);
-        }
-
-        public IEnumerator<object[]> GetEnumerator()
-        {
-            return data.GetEnumerator();
-        }
-
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return GetEnumerator();
-        }
-    }
-}
