+++ Backend.Test/Fixtures/CategoryFixture.cs
@@ -1,15 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace Backend.UnitTest;
 
-namespace Backend.Test
+public class CategoryFixture
 {
-    public class CategoryFixture
-    {
-        public CategoryManager CategoryManager { get; set; } = new CategoryManager();
+    public CategoryManager CategoryManager { get; set; } = new CategoryManager();
 
 
-    }
 }
