+++ Backend.Test/Fixtures/ItemFixture.cs
@@ -1,13 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace Backend.UnitTest;
 
-namespace Backend.Test
+public class ItemFixture
 {
-    public class ItemFixture
-    {
-        public ItemManager ItemManager { get; set; } = new ItemManager();
-    }
+    public ItemManager ItemManager { get; set; } = new ItemManager();
 }
