+++ Backend/Backend.Api.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>net7.0</TargetFramework>
@@ -13,6 +13,10 @@
 
 	<ItemGroup>
 		<InternalsVisibleTo Include="Backend.Test" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\Backend.Core\Backend.Core.csproj" />
 	</ItemGroup>
 
 </Project>
