--- Backend/Interfaces/IBudgetFiller.cs
@@ -1,9 +1,0 @@
-﻿namespace Backend.Interfaces
-{
-    public interface IBudgetFiller
-    {
-        public Budget? FetchLargeBudget();
-        public Budget? FetchMediumBudget();
-        public Budget? FetchSmallBudget();
-    }
-}
