--- Backend/Interfaces/IItem.cs
@@ -1,9 +1,0 @@
-﻿namespace Backend.Interfaces
-{
-    public interface IItem
-    {
-        public int Id { get; set; }
-        public string? Name { get; set; }
-        public float Amount { get; set; }
-    }
-}
