--- Backend/Interfaces/IItemManager.cs
@@ -1,9 +1,0 @@
-﻿namespace Backend.Interfaces
-{
-    public interface IItemManager
-    {
-        public bool CheckIfItemsAreValidInBudget(Budget budget);
-        public bool CheckValidItem(Item item);
-
-    }
-}
