--- Backend/Models/Budget.cs
@@ -1,12 +1,0 @@
-﻿using Backend.Interfaces;
-
-namespace Backend.Models
-{
-    public class Budget : IBudget
-    {
-        public Category Income { get; set; } = new Category();
-        public string? Title { get; set; }
-        public List<Category> Expenses { get; set; } = new List<Category>();
-        
-    }
-}
