--- Backend/Models/Item.cs
@@ -1,27 +1,0 @@
-﻿using Backend.Interfaces;
-using System.Globalization;
-using System.Runtime.CompilerServices;
-
-namespace Backend.Models
-{
-    public class Item : IItem
-    {
-        private static int IdChecker = 1;
-        public int Id { get; set; }
-        public string? Name { get; set; }
-        public float Amount { get; set; } = 0;
-
-        public Item()
-        {
-            if(Id == 0)
-            {
-                Id = IdChecker;
-                IdChecker++;
-            }
-        }        
-
-    }
-}
-
-
-
