+++ Backend/Program.cs
@@ -1,59 +1,58 @@
 
-using System.Security.Cryptography.X509Certificates;
 
-namespace Backend
+
+namespace Backend.Api;
+
+public class Program
 {
-    public class Program
+    public static void Main(string[] args)
     {
-        public static void Main(string[] args)
+        var builder = WebApplication.CreateBuilder(args);
+
+        // Add services to the container.
+        builder.Services.AddCors(options =>
         {
-            var builder = WebApplication.CreateBuilder(args);
-
-            // Add services to the container.
-            builder.Services.AddCors(options =>
+            options.AddPolicy("MyCorsPolicy", builder =>
             {
-                options.AddPolicy("MyCorsPolicy", builder =>
-                {
-                    builder.WithOrigins("http://localhost:5173/")
-                    .AllowAnyHeader()
-                    .AllowAnyMethod();
-                });
-
-                // Todo, fix new policy  ? 
-                options.AddPolicy(
-                    name: "AllowAllOrigins",
-                    policy =>
-                    {
-                        policy.AllowAnyHeader()
-                        .AllowAnyMethod()
-                        .AllowAnyOrigin();
-                    });
+                builder.WithOrigins("http://localhost:5173/")
+                .AllowAnyHeader()
+                .AllowAnyMethod();
             });
-            builder.Services.AddControllers();
-            // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-            builder.Services.AddEndpointsApiExplorer();
-            builder.Services.AddSwaggerGen();
 
-            var app = builder.Build();
+            // Todo, fix new policy  ? 
+            options.AddPolicy(
+                name: "AllowAllOrigins",
+                policy =>
+                {
+                    policy.AllowAnyHeader()
+                    .AllowAnyMethod()
+                    .AllowAnyOrigin();
+                });
+        });
+        builder.Services.AddControllers();
+        // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+        builder.Services.AddEndpointsApiExplorer();
+        builder.Services.AddSwaggerGen();
 
-            app.UseCors("AllowAllOrigins");
+        var app = builder.Build();
 
-            // Configure the HTTP request pipeline.
-            if (app.Environment.IsDevelopment())
-            {
-                app.UseSwagger();
-                app.UseSwaggerUI();
-            }
+        app.UseCors("AllowAllOrigins");
 
-            app.UseHttpsRedirection();
-
-            app.UseAuthorization();
-
-
-            app.MapControllers();
-
-            app.Run();
-
+        // Configure the HTTP request pipeline.
+        if (app.Environment.IsDevelopment())
+        {
+            app.UseSwagger();
+            app.UseSwaggerUI();
         }
+
+        app.UseHttpsRedirection();
+
+        app.UseAuthorization();
+
+
+        app.MapControllers();
+
+        app.Run();
+
     }
 }
